[grid-ui] Refactoring node block to smaller components
